Begin name refactor for V8Change
